use unsigned int as input to inttohex

+++ requests.go
@@ -155,6 +155,6 @@ func HexToInt(hex interface{}) uint64 {
 	return uint64(result)
 }
 
-func IntToHex(i int64) string {
-	return fmt.Sprintf("0x%s", strconv.FormatInt(i, 16))
+func IntToHex(i uint64) string {
+	return fmt.Sprintf("0x%s", strconv.FormatInt(int64(i), 16))
 }
